Setting the ALB to target prometheus, instead of Nginx
We did this in order to get a running prometheus for scraping purposes &
to ensure that we can access prometheus by LB URL. This is a temp measure
so that we can get something up & running. We had to do this because nginx
ran in to issues with service discovery. We can fix service discovery &
add nginx back in later.