move prometheus config to S3 bucket
This commit moves away from the terrible-testing-repo to an S3
bucket.  The way it works is:

 - there is a config bucket called gdsprom-$STACK-config
 - configuration files go in this bucket
 - tasks run with IAM roles which allow them to access appropriate
   parts of that bucket

There's still the matter of getting config from S3 onto the local
filesystem.  For this I'm still using the same pattern as the previous
git-puller task, but it's pulling from S3 instead of git.

This has some advantages:

 - our config is all in one source-controlled repo (ie, this one)
 - we use IAM roles to control fine-grained access to config files
 - we can actually update our config (the git-puller task only ever
   succeeded for initial clone, and never pulled in updates)

There are still some issues:

 - having two separate services to manage config like this is not a
   good pattern.  I'd rather have each task pull in its own config.

We might want to do things in future like:

 - baking our own docker images so that the container itself first
   grabs config from S3, then runs the binary (such as prometheus,
   nginx, alertmanager)
 - baking the config directly into our docker images?  (I'm not so
   keen on this as our config might change faster than we want to
   rebake images)