fold config grabbing task into prometheus task
This commit fold the former git-puller task into the prometheus task.
How it works:

 - when you have multiple containers in the same task definition, all
   tasks will run on the same instance. (This is what we want anyway)
 - if a task is marked as `essential: false`, then if it exits the
   task as a whole will continue working as normal.

This means that, when we run the prometheus task, two containers are
run:

 - s3-config-grabber, which pulls config down from S3 into /etc/prometheus
 - prometheus, which runs the binary using that config

s3-config-grabber will exit pretty quickly, but because it is
`essential: false`, the task as a whole continues.  In contrast, if
the prometheus container ever exits, the whole task (ie both
containers) will restart.

There's still a race condition between the config grabber and the
prometheus binary, though: we might well start the prometheus server
before the config has been grabbed, and prometheus will see the old
version of the config.