Revert "remove stream prefixes"
This reverts commit ca79ff54afb80482a2072396abe1d806d3da18cf.

It turns out that the stream prefixes have quite confusing behaviour.
If you *don't* specify one, your log stream is named with the docker
container id alone (something like
187c62b431e935d9c2e238744fc70a1477fe09ace505ea98af218b9bb3483cd1),
which you can only access by going on the box.

If you *do* specify one, your log stream is named in the format
`<prefix>/<container-name>/<ecs-task-id>`.  Note that the ECS task id
is not the same as the container id, and is probably more useful for
us since it's visible in the console and available via the AWS API.
So, even if there is no useful prefix to use, it's worth specifying
one anyway, so that you get container name and task id in the log
stream name.

In future, when we may have more than one container per task, it
becomes even more valuable to specify a prefix, because it groups
together the containers that are in the same task.