remove cpu values
The "cpu" parameter is optional and listed under advanced container
parameters in the ECS docs [1].  It only has an affect when CPU is
under contention (ie it doesn't constrain one container to use, say,
50% of the CPU if the other containers are idle).  We were also
setting it to some weird values.

It's tied to the underlying docker CpuShare contraint [2], which has a
default value of 1024.  I suspect if we leave this unset, everything
will get this default value.

If we need to tweak this in future, we can always revisit.

[1] https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#advanced_container_definition_params
[2] https://docs.docker.com/engine/reference/run/#cpu-share-constraint