Use memoryReservation instead of memory for prometheus
We have an issue that we're unable to redeploy prometheus because of
our memory reservation parameters.  The issue is:

 - the container instance has slightly less than 16G available
 - the prometheus task reserves 8G for prometheus and 256M for the
   config grabber
 - this means that two concurrent task instances would consume more
   memory than the container instance has
 - so the new one can't be deployed and we get stuck

To solve this, we replace `memory` with `memoryReservation`.  They are
documented in the ECS docs [1], but in summary:

 - `memory` specifies the hard limit of how much memory the container
   can use
 - `memoryReservation` specifies how much memory will be reserved for
   the container's exclusive use, but the container can use memory
   beyond this limit if it is available

You must specify at least one of `memory` and `memoryReservation`.  If
you only specify `memory`, then `memoryReservation` will have the same
value.  If you only specify `memoryReservation`, then `memory` will be
effectively infinite (the container can use all available memory).  If
you specify both, then `memoryReservation` must be less than
`memory`.

In our case, we want prometheus to have a minimum of 4G available (so
we specify `memoryReservation` of 4G) but we also want it to be able
to use all available memory if needed (so we don't specify `memory`).

For the s3-config-grabber, the existing value of `memory` set to 256M
seems sensible.

[1] https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#standard_container_definition_params