Add travis CI with check for `terraform fmt`
This is copied from
https://github.com/alphagov/govuk-aws/blob/master/tools/terraform-format.sh

The script is set up in a way to support git pre commit hooks
(passed a list of files to be checked) rather than being written
like `terraform fmt` which would naturally iterate over all files.

We can later add pre commit hooks for ease of use.

Choice of Travis has been made in terms of speed of set up but
can come with some drawbacks such as having to wait on Travis
when America wakes up. It should be sufficient for the near future.