Add a Jumpbox to dev stacks
Allows developers to access the first EC2 instance via a jumpbox in order to investigate any issues with apps running if there are insufficient logs in AWS.
This should work even if the IP of the EC2 instance changes, as the fingerprint is removed from known_hosts before it is added back on when sshing onto the jump and EC2 instance.

The jumpbox will not be available for non dev environments, i.e. staging and production.

In order to prevent the jumpbox from being destroyed with each apply, the jumpbox module had to ignore changes to 'security_groups', so if the 'infra-security-groups' project does change, the jumpbox project will need to be tainted in order to pick up the updates.

It creates an ssh key based on the users public key (it assumes the users public key is available as `~/.ssh/id_rsa.pub`).
The jump box adds ingress rules for port 22 to `monitoring_external_sg`.

You should set the `ecs_instance_ssh_keyname` variable to the name of the ssh pair generated by the ssh_key module. This will enable to ssh agent forwarding and log onto the ecs instance. It's a good idea to remove the tfvars for your stack and regenerate it using `make create-stack`.

Note: Changing the ssh key requires the 'app-ecs-instances' project to be reapplied.