Add pagerduty integration
This adds integration with a re-tools team wide pager duty account. Alertmanager will send an alert to pagerduty when an alert is triggered on our staging or production prometheis.
Alerts will not be triggered on development stacks. Staging and production Prometheis are sharing a pagerduty service, this was done for speed but can be changed later if we need a lager degree of separation.

https://trello.com/c/dljFeMNm/426-integrate-alertmanger-with-pagerduty-to-enable-alerts-to-become-pages