Make route 53 zone id available as an output
We want to move the route 53 alias record from this project to live
next to the ALB that it is referencing.  To do this, we need to expose
the zone id as a project output, which will make it available in the
statefile for consumption by app-ecs-albs.

This commit merges the dev and staging/prod route 53 zone resources
together, so that there is only a single zone id to make available in
the outputs.

When running this change, the dev zones will be destroyed and
recreated.  We feel that's okay.