Add HTTPS to prometheus and alertmanager
The changes here enable SSL encryption for your service endpoints.
A certificate is issued & stored in AWS certificate manager. The
ALB can then make use of the certificate. All incoming requests
to the ALB to port 443 will be served HTTPS traffic. It is still
possible to use the HTTP endpoint, however we can drop that
support later on when we are happy with HTTPS is working fine.

Validation of the certificate is done using DNS validation. Note,
that as DNS validation can take a little while (anywhere from a
couple of minutes up to half an hour) there is a chance that dev
stacks will take much longer to create. It is likely suggested that
devs don't tear down their stack fully and instead keep a dev stack
for themselves which they make changes to.

Certificate renewal is handled by AWS automatically and if this
fails then we will get an email sent to the address associated
with the AWS account.