Fix updating bucket objects bug
This commit:

- Removes previous bucket policy
- Adds automatic encryption to the bucket
- Readds etags to our bucket objects

This fixes a bug whereby changing an object, such as alerts.yml,
would not be picked up by Terraform and would therefore not be
updated when running a Terraform apply.

It also ends up being a nicer solution with less lines of code
to ensure all objects in our config bucket are encrypted at rest.