Deploy container instances in separate ASGs
This commit splits the one-ASG-with-three-instances into
three-ASGs-with-one-instance-each.

The goal I am going for with this commit is: we need to be able to
tell each separate prometheus service
aws_ecs_service.prometheus_server to run on a separate instance.  I
would also like to make our code flexible enough that dev environments
can run one single instance, and staging and production could scale
down to 2, or up to 4 (with two prometheis in a single AZ), if desired.

I had considered telling the services to run with placement contraints
[1] pinning each prometheus service to a specific availability zone.
However, if there are only 1 or 2 instances, then some of the AZs will
be empty, and pinning a service to that AZ will mean it can never run
at all.  Moreover, there's no clean way to know which AZs are
available.  Finally: if an instance in an ASG dies, will the instance
that replaces it be in the same AZ, or could it come back in a
different one?

I thought a better way of doing this would be to set an explicit
placement attribute in the ECS agent for each instance.  In order to
do this, each instance needs to be created by a separate ASG (because
an ASG has a single user data script, and the user data script sets up
the ECS agent settings).  Now we can pin prometheus services to
specific instances using the custom `prometheus-instance` attribute,
and have confidence that those specific instances exist.

Unfortunately, terraform modules don't currently support the count
attribute [2], so the only way to have multiple ASGs using this module
is to copy/paste the module multiple times.  This feels like the
least-worst option to me.  (In future, to scale up we would need to
copy/paste more.  To make the number of instances configurable, we can
make each ASG conditional based on the desired number of instances.
For details, see the module docs [3].)

This has a side benefit that we can now specify the exact volume id in
user data again, rather than doing the trick where we grab a list of
volumes from the API and select the one that's in the same AZ as
us.  (This also enables us to deploy multiple instances in the same AZ
in future if we want.)

In order to be certain that we're pairing up availability zones,
private subnets, ASGs and volume ids correctly, I introduced a
data.aws_subnet.private_subnets data source, and pull availability
zone config from there to configure the aws_ebs_volumes.  This means
we can be confident that the EBS volume is in the same AZ as the ASG
it's paired with.

Note that the number 3 is still hardcoded throughout
app-ecs-instances/main.tf; actually making things configurable will
come in future commits.

[1]: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-placement-constraints.html
[2]: https://github.com/hashicorp/terraform/issues/953
[3]: https://registry.terraform.io/modules/terraform-aws-modules/autoscaling/aws/2.6.0