drive EBS and ASG placement from same source
Previously, we distributed EBS volumes across AZs and configured the
ASG with a list of subnets, but there was nothing in the
app-ecs-instances project that ensured any alignment between the
particular AZs used by the subnets and by the EBS volumes.  This isn't
a problem if the number of instances/EBS volumes equals the number of
AZs, but if we want dev environments to have fewer instances, we
should be able to point to code that demonstrates that the subnets in
the ASG are the same as the AZs in the EBS volumes.