make number of instances configurable
This makes the number of instances configurable by a terraform
variable prometheis_total.  It can have any value from 0 to 3.  It
will control the number of instances created in the autoscaling group,
and in turn the number of prometheus/alertmanager servers created.

The tricky part here is ensuring that we assign services to
availability zones which actually have instances in them.  If we just
let the autoscaling group have all subnets (and hence all availability
zones), I don't know what guarantees there are about which AZs it will
choose to use.

In this commit, we ensure that the ASG only has access to a subset of
subnets, that the EBS volumes are created in AZs corresponding to
those subnets, and that the services are assigned to those subnets.
This way, we avoid a possible situation where an instance is created
in an AZ which has no EBS volumes (and vice versa).

After this commit, you can add

prometheis_total = 1

to your stacks/foo.tfvars file in order to reduce the number of
instances you provision.