Extract nginx auth proxy into own ECS service
This commit extracts nginx auth proxy into it it's own service splitting
it away from Prometheus.

Before this commit Prometheus and nginx were co-located in the same
service, nginx communicated with Prometheus using a container link, this
is inconsistent with how nginx communicates with Alert Manager and
because of the basic auth prevents Prometheus from scrapping the other
instances of Prometheus.

New Nignx Service and Task
==========================

This commit defines a new service in `nginx-auth-proxy.tf` which
implements the new `nginx-auth-proxy.json` task.
The nginx auth has been renamed and updated to point to the new
internal dns names of prometheus.

Changes to ALBs
===============

A few resources have been renamed to be more consistent, clearer. We
have been able to remove the listener rules for the external dns. When
Prometheus and Nginx where co-located we needed to route requests to the
appropriate instance of Prometheus, which meant we needed to route the
request to the specific Nginx that Prometheus was paired with. Now we
can route to any Nginx each Nginx knows how to route a request to any
Prometheus. This means we can default to sending all requests to a
single Nginx target group which all three instances of Nginx are
registered to and let Nginx manage the routing to the instances of
Prometheus.

A new set of internal DNS names have been setup and a new set of
listener rules have been set up for routing Prometheus traffic to using
the internal load balancer.

Changes to Prometheus Service and Task
======================================

The old Nginx task has been removed and the new service now targets the
Prometheus container

Changes to Nginx Config
=======================

There is now a default server for responding to health check requests.
There are 3 new servers for responding to the Prometheus checks. If a
request is made to the load balancer directly it should 404 (unless it
is `/health`) at the default server.