Change registers alert
The previous alert was really quite spammy.  We sat down and talked
about what might be a better measure of failure.

For context:

 - a lot of 5xx errors on the registers app are caused by bots making
   weird requests; they therefore don't really result in real users
   being impacted and the alert isn't actionable
 - the failure modes we are interested in are things like "we broke
   the app and now basically every request fails (except maybe things
   like /favicon.ico)

As a result, we decided to change the alert to be "some really high
percentage of all requests are erroring".

This alert says "if the total requests over a 5 minute rolling window
has more than 25% 5xx errors, and continues to do so for 2 minutes
solidly, raise an alert".