Add an ASG schedule for dev stacks
The schedule is based on UTC times.
The default list of cron schedules is based around the core hours of 9 - 18 hours, Monday to Friday.
As developers can spin up the instances outside of these core hours to ensure that the EC2 instances are terminated the scheduled EC2 termination will be scheduled for at each hour outside of the core hours.
If you have a different working pattern, you can add the `asg_dev_scaledown_schedules` variable in your tfvars files with a list of cron schedules.