Update setup.sh and prometheus task definition
Updated setup.sh to remove the tfvars file when doing a clean of the stack, this is to ensure that when creating new dev stacks it uses the latest variables. In this case rather than using m4.xlarge use t2.small, and to pick up lower prometheus cpu and memoryReservation values for dev stacks.

The default stack tfvars has been updated with default dev values which can be overridden by devs.

The prometheus task definition was changed in order to use the variables from tfvars for your stack if they exist otherwise default values will be used which are the values for production and staging.