Do not rename the external alb
The external alb was renamed as a result of creating a new alb alongside
the old alb then switching traffic to the new one. The old alb was then
removed. This resulted in the alb getting rebuilt and DNS updated. The
rebuild of the ALB took over 2 minutes and the changes to the DNS can
result also result in down time due to DNS caching issues.

This commit reverts the terraform name of the ALB, this means that the
ALB is not rebuild and the DNS does not change.

There is still some downtime whilst the app is rebuilt, this results in
503 in the ALB, but the amount of downtime is much reduced.