Only trigger registers alert in production PaaS space
Registers saw an alert trigger yesterday but it turned out
they did not want it to trigger because it was for a non production
app that they were running in their `sandbox` PaaS space.

This commit means that only apps in their `prod` PaaS space are
alerted on.

Note, it appears that because there is an label overwrite as
per https://prometheus.io/docs/prometheus/latest/configuration/configuration/
(see `honor_labels` on that page), we have to use the `exported_space`
rather than `space`. I will raise a separate trello ticket to
see if we can make these labels clearer/more user friendly.