Mesh alertmanagers to prevent alert email duplication
The purpose of this PR was to enable alert manager meshing so that the
email are not duplicated when alerts trigger.

This is based on the following card: https://trello.com/c/Fpir7Ilk/509-enable-prometheus-alerts-to-be-delivered-by-email

I had to change the config file to the default name since I had issue's with defining commands.
By not specifying the alertmanager config I was able to just pass the list of mesh target to the task definition.