Removing security rules that are no longer required
We use to run prometheus and alert manager in fornt of an ALB
with an Nginx server with each service. There where multiple targets
groups which exposed alertmanager to the external facing ALB. This
meant that we needed alot of security group rules in order to
permit access to external facing ALB from the public. We also needed
internal security groups to ensure that communciation was occuring.

We have now added a signle nginx service that enables routing based
host header, including similar rules to ALB's. This means
we now nolonger need to many rules.