Mesh alertmanagers to prevent alert email duplication
The purpose of this PR was to enable alert manager meshing so that the
email are not duplicated when alerts trigger.

This is based on the following card: https://trello.com/c/Fpir7Ilk/509-enable-prometheus-alerts-to-be-delivered-by-email

There is a few unexpected things occuring. We ended up deciding to go with nginx reverse proxy
in order to perform the meshing. We used DNS names that auto update at instance start up time.
This ensured us that the instance the IP address will not go stale.