simplify alerting rules check script
Bash will propagate the exit code of the last command executed, so we
don't actually need a conditional here.

We also don't need to add any output to what `promtool check rules`
already reported.

For safety though, it's worth adding `set -e` so that bash will
immediately exit if any command in the file fails.  In this case, it
doesn't matter because there's only one command so we exit afterwards
anyway, but it makes the file robust against future edits.