remove basic auth requirement for office IPs
This updates our nginx config to allow users from the office IPs to
bypass the basic auth requirement.  Users outside the office IPs can
still access prometheus if they do have the basic auth creds.  This
means that, in particular, grafana can access prometheus.

The pieces of this commit are:

 - cidr_admin_whitelist: the office IP addresses - taken from internal IT wiki
 - `satisfy any`: this nginx directive says "if the user has any of:
   an approved IP address or basic auth creds, let them in"
 - `real_ip_header` and `set_real_ip_from`: these get the real client
   IP from a header provided by the load balancer
 - `allow %s`: this is the directive (provided in the `formatlist`
   call) that allows access from a CIDR
 - `deny all`: if you haven't been allowed by any previous directive,
   you get denied access now