Include env files for staging and prod environments
Deployment has been slowed as these variables have not been known
or takes a bit of time to dig them out from elsewhere. This should
make things quicker.

Note, `PROFILE_NAME` is likely to differ by each developer as it
is the name you deceided on for your AWS vault profiles. I've
included mine for the moment which is likely not ideal but
may be an improvement to start with.