Enable make file to allow access to terraform console
Edit the terraform make file to allow us to use terraforms console feature.
This feature allows you to test interpolations and terraform funcations. This
will enable to us to reduce the number of times we run plan and apply. We can
test all our variables against our terraform funcations local before making an apply.

This is one of the thing that slowed down development time for me.
This commit will greatly improve the development process for me and hopefully
for the team also.