Remove multi nat-gateways in dev to save on cost
The dev enviroment currently creates three of every network component  based
on the configuration in the vpc module. I have added a small change which will
prevent the dev enviroment from creating a nat gateway for each AZ. If the
instance count is 3 in dev the single nat gateway will still be enabled. This
is fine since the dev enviroment is not HA enviroment. There will be extra routes
created to route traffic through the AZ which has all the active nat-gateway.

The current production enviroment currently supports an gateway per AZ. This is for HA
purposes.