Add PR template
This adds a file which will be used to pre-populate new pull
requests.  I thought it's something we could try to make our PRs
easier to review.

In order to review a PR, you need to know what the author was trying
to do in order to judge how effectively the PR achieves the stated
goals.  Therefore, the initial template has two sections: "Why I am
making this change" and "What approach I took".

This is very much a work in progress and we can iterate it as we go to
see what works.

https://blog.github.com/2016-02-17-issue-and-pull-request-templates/