Allow alertmanager mesh to prevent duplicate states
This will allow alertmanager's to mesh with each other in order to
prevent duplicate states. There is a few changes that include creating a new task definition.
This task definition will permit alert managers to mesh and it will also provide a single place
to provide arguments for the task definition. I tried to make it as clean as possible. Please feel
free to change anything you would like.