Create Prometheus Enclave
This commit deploys prometheus.

There are two modules defined, enclave/network and enclave/prometheus

There are two projects enclave/verify-perf-a/network and
enclave/verify-perf-a/prometheus

The projects instantiate the respective modules with variables for that
particular environment, for example the vpc_id which verify-perf-a
specific is defined in projects/enclave/verify-perf-a/network and passed
through to the network module which can then use the vpc id to build the
subnets.

Using this pattern it is easy to deploy prometheus to new verify
environments by creating new enclave projects and specifying the
appropriate variables.

The network module defines two subnets in different availability zones,
whist not strictly necessary (verify only operates out of a single
availability zone) it was felt it was easier to do it now then build
with a single instance and add the second later.

The network also describes the security groups and vpc endpoints. VPC
endpoints are used to reach s3 and the ec2 api. The s3 endpoint is an
example of a "Gateway" endpoint, it makes external IP address routeable
without going over the public internet. The ec2 endpoint is an example
of interface endpoint, again the goal is to access the ec2 api endpoint
without needing to access the public internet. However in this case
instead of using routing rules to reach the endpoint a network interface
is added to the subnet. If you want to reach the service you connect to
the local ip of the network interface instead of the public ip.
This caused a problem, prometheus uses the EC2 API for service
discovery, prometheus uses the aws go sdk which does not allow you to
override the ip address used to connect to the service. The sdk will
always try and connect via ec2.eu-west-2.amazonaws.com (the sdk does let
you configure proxies but that doesn't help in this case) we therefore
have to override the address of ec2.eu-west-2.amazonaws.com in the hosts
file. To do this required a data lookup against the network interfaces
and a bit of terraform trickery to pull out a list of private_ips for
the network interfaces. This is a little hacky so the ec2-endpoint code
has been kept in a separate file to make it easier to remove if decided
to adopt a different approach.

The prometheus module defines the prometheus istances, the disks and the
prometheus config. The instances are built using cloud init. The config
file is published to an s3 bucket. The instances poll the s3 bucket and
download the config and restart the prometheus service. Currently the
state of the scripts which manage the process are proof of concept, they
work but with caveats, currently prometheus is restarted every 2 minutes
because the scripts are not able to tell if the config has been modified
and there is no validation of the config. Prometheus reloads config
without interruption to the service and promethues ignores invalid
config and continues running with the old. So whilst the config updating
process could be improved the impact of the existing scripts is minimal.

Currently the configuration assumes it is being ran within a verify
environment this was to speed up development.

A deploy script was created to help with the running of terraform. It is
proof of concept quality too. An example command `./deploy_enclave.sh -e
verify-perf-a -p gds-tech-ops -a apply -s prometheus` will apply
the prometheus terraform the verfiy-perf-a enclave. The script assumes
role to the promethues_deployer role in the verify account, changes
directory in to correct project directory and executes terraform -a is
the action so you can `-a apply` to apply `-a plan` to run a plan and
`-a destroy` to destroy the environment. The apply runs a plan first
saving the plan file to be used in the apply currently the script does
not clean up the plan files.