Use `nginx:stable-alpine` rather than `nginx:alpine` images
Deployments are currently failing due to
https://github.com/nginxinc/docker-nginx/issues/230

If an instance was to go down then nginx containers could not start
and our service would go down.

This is a hotfix that should allow nginx tasks to restart if needbe
as suggested by other users in the github issue by swapping
to an older nginx image that can be found.

We may want to address pinning to an exact version at a later point
but this should do as a hotfix.