initial test kitchen setup
Adding Basic IaC test to prometheus enclave enviroment

We need to ensure that the we are able to test the funcationality of the prometheus
enviroment. We use kitchen an Inspec to achieve this. This commit add this to current infratsture tests.
We spin up an enviroment completely independent of the current set up.

initial test kitchen setup

Adding Basic IaC test to prometheus enclave enviroment

We need to ensure that the we are able to test the funcationality of the prometheus
enviroment. We use kitchen an Inspec to achieve this. This commit add this to current infratsture tests.
We spin up an enviroment completely independent of the current set up.