Add terraform projects for paas staging and production
Add security group to allow ssh access to enclave prometheus instance.

- CIDRs are office IPs
- default had to be introduced for ec2_endpoint_ips in order to provide a value to the aws_ec2_ip var which expects a value but it's not needed for PaaS discovery.

Add terraform for enclave prometheus staging and production environments.
The config bucket created for the stacks are created in eu-west-2 as the creation of the bucket is defined in deploy_enclave.sh.

Create verify-config module moving prometheus config to s3 bucket.

- verify-perf-a stack, move config to verify-config module and added prometheus.conf.tpl
- Attempting to add data to the existing prometheus-staging bucket causes problems, it can't find the s3 resource.
- Add SG for port 9090 to enable access to metrics for Prometheus

Create paas-config module to allow for the prometheus config to be created and uploaded to the prometheus EC2 instance using cloud.conf.

Fixed typos in verify perf a prometheus terraform, this will cause EBS volumes to be replaced.