change IP of egress proxies
10.0.1.87 was the IP of egress-proxy-1.mgmt.dmz.  However we have
changed our configuration so that we go via egress-proxy.service.dmz
which is a DNS round-robin to two amazon load balancers which in turn
balance to egress-proxy-{1,2}.mgmt.dmz.

The load balancers are on dynamic IPs so these numbers are liable to
change.  Hardcoding these values should be considered technical debt.