Add alert for Sidekiq queues for production DATA.gov.uk
For our production instance, if we have more than 800 enqueued jobs for at least 5
minutes, raise an alert. Looking at the historical jobs for sidekiq, it
seems like we get spikes up to around 1.2k jobs. The spikes clear down
in less than 5 minutes. 800 jobs is just over 65% of the max spike we
get so it seems like a reasonable point to alert, especially since we
expect all the jobs to have processed in under 5 minutes.