Allow Prometheus to be accessed via Verify jumpbox
Verify use an nginx reverse proxy to expose their tools, graphite,
sensu etc. They have asked us to expose prometheus in the same way.

The jump box must be able to reach the prometheus instances, we have
created an `allow_web_interface` ingress rule to allow this.

Verify need to know the ip address of the Prometheus instances to be
able to direct traffic to them. Prometheus IP addresses are not hard
coded and can change when instances are rebuilt. We therefore cannot
hard code the IP address in the verify config.
Instead we need to use DNS, verify have a private hosted zone which
we can use to add the addresses of the promethues instances. This code
is encapsulated as a module to enable us to only included it in the
verify project.